Separando main.dart em diferentes arquivos para organizar melhor o projeto
