+++ my_app/lib/models/exercise_model.dart
@@ -1,0 +1,6 @@
+class Exercise {
+  final String name;
+  final String series;
+
+  const Exercise({required this.name, required this.series});
+}
