+++ my_app/lib/models/readme
@@ -1,0 +1,1 @@
+Estes arquivos definem a estrutura dos seus dados (classes).
