+++ my_app/lib/models/user_model.dart
@@ -1,0 +1,19 @@
+class User {
+  String name;
+  String email;
+  int height;
+  int weight;
+  int age;
+  String objective;
+  String profileImageUrl;
+
+  User({
+    required this.name,
+    required this.email,
+    required this.height,
+    required this.weight,
+    required this.age,
+    required this.objective,
+    required this.profileImageUrl,
+  });
+}
