+++ my_app/lib/screens/edit_profile_page.dart
@@ -1,0 +1,77 @@
+import 'package:flutter/material.dart';
+import '../models/user_model.dart';
+
+class EditProfilePage extends StatefulWidget {
+  final User user;
+  const EditProfilePage({Key? key, required this.user}) : super(key: key);
+
+  @override
+  _EditProfilePageState createState() => _EditProfilePageState();
+}
+
+class _EditProfilePageState extends State<EditProfilePage> {
+  late TextEditingController _nameController;
+  late TextEditingController _heightController;
+  late TextEditingController _weightController;
+  late TextEditingController _ageController;
+  late TextEditingController _objectiveController;
+
+  @override
+  void initState() {
+    super.initState();
+    _nameController = TextEditingController(text: widget.user.name);
+    _heightController = TextEditingController(text: widget.user.height.toString());
+    _weightController = TextEditingController(text: widget.user.weight.toString());
+    _ageController = TextEditingController(text: widget.user.age.toString());
+    _objectiveController = TextEditingController(text: widget.user.objective);
+  }
+
+  @override
+  void dispose() {
+    _nameController.dispose();
+    _heightController.dispose();
+    _weightController.dispose();
+    _ageController.dispose();
+    _objectiveController.dispose();
+    super.dispose();
+  }
+
+  void _saveProfile() {
+    final updatedUser = User(
+      name: _nameController.text,
+      email: widget.user.email,
+      height: int.tryParse(_heightController.text) ?? widget.user.height,
+      weight: int.tryParse(_weightController.text) ?? widget.user.weight,
+      age: int.tryParse(_ageController.text) ?? widget.user.age,
+      objective: _objectiveController.text,
+      profileImageUrl: widget.user.profileImageUrl,
+    );
+    Navigator.pop(context, updatedUser);
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text('Editar Perfil'),
+        actions: [IconButton(icon: const Icon(Icons.save), onPressed: _saveProfile)],
+      ),
+      body: ListView(
+        padding: const EdgeInsets.all(16.0),
+        children: <Widget>[
+          TextField(controller: _nameController, decoration: const InputDecoration(labelText: 'Nome')),
+          const SizedBox(height: 16),
+          TextField(controller: _heightController, decoration: const InputDecoration(labelText: 'Altura (cm)'), keyboardType: TextInputType.number),
+          const SizedBox(height: 16),
+          TextField(controller: _weightController, decoration: const InputDecoration(labelText: 'Peso (kg)'), keyboardType: TextInputType.number),
+          const SizedBox(height: 16),
+          TextField(controller: _ageController, decoration: const InputDecoration(labelText: 'Idade'), keyboardType: TextInputType.number),
+          const SizedBox(height: 16),
+          TextField(controller: _objectiveController, decoration: const InputDecoration(labelText: 'Objetivo')),
+          const SizedBox(height: 32),
+          ElevatedButton(child: const Text('Salvar Alterações'), onPressed: _saveProfile),
+        ],
+      ),
+    );
+  }
+}
