+++ my_app/lib/services/readme
@@ -1,0 +1,1 @@
+Este arquivo isola a lógica de salvar e carregar os dados do usuário.
